requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.6
django-recaptcha==2.0.6
django-utils-six==2.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
